+++ HISTORY.md
@@ -1,3 +1,9 @@
+v3.15.6
+===================
+## Bug Fixes:
+*  Remove mnokey patch `px-map-layer-geojson` close popup when data changes
+*  Fix: `px-map-layer-geojson` close popup when data changes
+
 v3.15.5
 ===================
 ## Bug Fixes: